refactor vdj
