--- celescope/templates/html/vdj_blast/base.html
@@ -1,14 +1,0 @@
-
-{% include "html/utils/start.html"%}
-
-{% if consensus_summary is defined %}
-{% set step_summary = consensus_summary %}
-{% include "html/utils/step.html"%}
-{% endif %}
-
-{% if mapping_vdj_summary is defined %}
-{% set step_summary = mapping_vdj_summary %}
-{% include "html/utils/step.html"%}
-{% endif %}
-
-{% include "html/utils/end.html"%}
