+++ celescope/vdj/__init__.py
@@ -1,5 +1,6 @@
-__STEPS__ = ['sample', 'barcode', 'cutadapt', 'consensus', 'mapping_vdj', 'count_vdj']
+__STEPS__ = ["sample", "barcode", "cutadapt", "consensus", "mapping_vdj", "count_vdj"]
 __ASSAY__ = 'vdj'
+
 CHAINS = {
     "TCR": ["TRA", "TRB"],
     "BCR": ["IGH", "IGL", "IGK"],
