--- celescope/vdj_blast/__init__.py
@@ -1,6 +1,0 @@
-__STEPS__ = ["sample", "barcode", "cutadapt", "consensus", "mapping_vdj"]
-__ASSAY__ = 'vdj_blast'
-CHAINS = {
-    "TCR": ["TRA", "TRB"],
-    "BCR": ["IGH", "IGL", "IGK"],
-}
